% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbxClusterAdmin.R
\name{autoscale}
\alias{autoscale}
\title{Returns an autoscale structure}
\usage{
autoscale(min_workers, max_workers)
}
\arguments{
\item{min_workers}{The minimum number of workers to which the cluster can scale down when underutilized. It is also the initial number of workers the cluster will have after creation.}

\item{max_workers}{The maximum number of workers}
}
\value{
an object of class autoscale
}
\description{
Returns an autoscale structure
}
\details{
see https://docs.databricks.com/api/latest/clusters.html#clusterautoscale
}

