% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitr_engines.R
\name{databricksPythonEngine}
\alias{databricksPythonEngine}
\title{Databricks Python Engine for Knitr}
\usage{
databricksPythonEngine(options)
}
\description{
Databricks Python Engine for Knitr
}

