% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/languageContext.R
\name{dbxCmdStatus}
\alias{dbxCmdStatus}
\title{Command Status}
\usage{
dbxCmdStatus(cmdId = getOption("databricks")$currentCommand,
  ctx = getOption("dbpycontext"),
  instance = options("databricks")[[1]]$instance,
  clusterId = options("databricks")[[1]]$clusterId,
  user = options("databricks")[[1]]$user,
  password = options("databricks")[[1]]$password)
}
\arguments{
\item{cmdId}{commandId returned by runCommand (by default, the current command)}

\item{ctx}{the context for the language (current python context)}

\item{instance}{is the instance of databricks}

\item{clusterId}{is the clusterId you're working with}

\item{user}{your usename}

\item{password}{your password}
}
\value{
a commandId
}
\description{
Command Status
}
\details{
see https://docs.databricks.com/api/1.2/index.html#command-execution
}

