% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/languageContext.R
\name{dbxCtxMake}
\alias{dbxCtxMake}
\title{Create a Remote Language Context}
\usage{
dbxCtxMake(language = "python",
  instance = options("databricks")[[1]]$instance, wait = TRUE,
  verbose = FALSE, clusterId = options("databricks")[[1]]$clusterId,
  user = options("databricks")[[1]]$user,
  password = options("databricks")[[1]]$password)
}
\arguments{
\item{instance}{is the instance of databricks}

\item{wait}{should i wait for context to start?}

\item{clusterId}{is the clusterId you're working with}

\item{user}{your usename}

\item{password}{your password}

\item{the}{language context}

\item{private}{should it modify options?}
}
\value{
a context object
}
\description{
Create a Remote Language Context
}
\details{
see https://docs.databricks.com/api/1.2/index.html#execution-context
}

