% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageContext.R
\name{dbxCtxStatus}
\alias{dbxCtxStatus}
\title{Get Status of a Context}
\usage{
dbxCtxStatus(ctx, instance = options("databricks")[[1]]$instance,
  clusterId = options("databricks")[[1]]$clusterId,
  user = options("databricks")[[1]]$user,
  password = options("databricks")[[1]]$password)
}
\arguments{
\item{ctx}{the context for the language}

\item{instance}{is the instance of databricks}

\item{clusterId}{is the clusterId you're working with}

\item{user}{your usename}

\item{password}{your password}
}
\value{
a context object
}
\description{
Get Status of a Context
}
\details{
see https://docs.databricks.com/api/1.2/index.html#execution-context
}
