% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbxClusterAdmin.R
\name{dbxDelete}
\alias{dbxDelete}
\title{Delete a cluster}
\usage{
dbxDelete(cluster_id, token = options("databricks")[[1]]$token,
  instance = options("databricks")[[1]]$instance)

dbxDelete(cluster_id, token = options("databricks")[[1]]$token,
  instance = options("databricks")[[1]]$instance)
}
\arguments{
\item{cluster_id}{a string}

\item{cluster_id}{a string}
}
\value{
nothing

a big json
}
\description{
Delete a cluster

Get information about  a cluster
}
\details{
see https://docs.databricks.com/api/latest/clusters.html#delete

see https://docs.databricks.com/api/latest/clusters.html#delete
}

