% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbxClusterAdmin.R
\name{dbxNodeTypes}
\alias{dbxNodeTypes}
\title{Get the different node types}
\usage{
dbxNodeTypes(token = options("databricks")[[1]]$token,
  instance = options("databricks")[[1]]$instance)
}
\arguments{
\item{token}{obtained from databricks}

\item{instance}{from frank}
}
\value{
a list with default node type and data table of nodes
}
\description{
Get the different node types
}
\details{
seehttps://docs.databricks.com/api/latest/clusters.html#list-node-types
}
