% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbxClusterAdmin.R
\name{dbxResize}
\alias{dbxResize}
\title{Resize a Spark cluster given its id and num_workers}
\usage{
dbxResize(cluster_id, token = options("databricks")[[1]]$token,
  instance = options("databricks")[[1]]$instance)
}
\arguments{
\item{cluster_id}{a string}

\item{num_workers}{an integer or of class autoscale}
}
\value{
nothing
}
\description{
Resize a Spark cluster given its id and num_workers
}
\details{
see https://docs.databricks.com/api/latest/clusters.html#resize
}

