% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbxClusterAdmin.R
\name{dbxRestart}
\alias{dbxRestart}
\title{Restart a cluster}
\usage{
dbxRestart(cluster_id, token = options("databricks")[[1]]$token,
  instance = options("databricks")[[1]]$instance)
}
\arguments{
\item{cluster_id}{a string}
}
\value{
nothing
}
\description{
Restart a cluster
}
\details{
see https://docs.databricks.com/api/latest/clusters.html#restart
}

