% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageContext.R
\name{dbxRunCommand}
\alias{dbxRunCommand}
\title{Command execute}
\usage{
dbxRunCommand(command, wait = 0, ctx,
  instance = options("databricks")[[1]]$instance,
  clusterId = options("databricks")[[1]]$clusterId,
  user = options("databricks")[[1]]$user,
  password = options("databricks")[[1]]$password)
}
\arguments{
\item{command}{a string command to be executed}

\item{wait}{if non zero, will wait till command is finnished. wait is seconds polling}

\item{ctx}{the context for the language}

\item{instance}{is the instance of databricks}

\item{clusterId}{is the clusterId you're working with}

\item{user}{your usename}

\item{password}{your password}
}
\value{
a commandId
}
\description{
Command execute
}
\details{
see https://docs.databricks.com/api/1.2/index.html#command-execution
}
