% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbxClusterAdmin.R
\name{dbxSparkVersions}
\alias{dbxSparkVersions}
\title{Get the different versions of spark available}
\usage{
dbxSparkVersions(token = options("databricks")[[1]]$token,
  instance = options("databricks")[[1]]$instance)
}
\arguments{
\item{token}{obtained from databricks}

\item{instance}{from frank}
}
\description{
Get the different versions of spark available
}
\details{
see https://docs.databricks.com/api/latest/clusters.html#spark-versions
}
