% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbxClusterAdmin.R
\name{dbxStart}
\alias{dbxStart}
\title{Starts a terminated Spark cluster given its id}
\usage{
dbxStart(cluster_id, wait = TRUE, token = options("databricks")[[1]]$token,
  instance = options("databricks")[[1]]$instance)
}
\arguments{
\item{cluster_id}{a string}

\item{wait}{if TRUE, wait till cluster starts}
}
\value{
nothing
}
\description{
Starts a terminated Spark cluster given its id
}
\details{
see https://docs.databricks.com/api/latest/clusters.html#restart
}

