% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbxClusterAdmin.R
\name{dbxZones}
\alias{dbxZones}
\title{Get the different zones}
\usage{
dbxZones(token = options("databricks")[[1]]$token,
  instance = options("databricks")[[1]]$instance)
}
\arguments{
\item{token}{obtained from databricks}

\item{instance}{from frank}
}
\value{
a list of default zone and data table of zones
}
\description{
Get the different zones
}
\details{
see https://docs.databricks.com/api/latest/clusters.html#list-zones
}

