% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/languageContext.R
\name{isCommandFinished}
\alias{isCommandFinished}
\title{check if command is finished}
\usage{
isCommandFinished(s)
}
\arguments{
\item{s}{a command is}
}
\description{
check if command is finished
}

