% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/languageContext.R
\name{isCommandRunning}
\alias{isCommandRunning}
\title{check if context is running}
\usage{
isCommandRunning(s)
}
\arguments{
\item{s}{a commandid}
}
\description{
check if context is running
}

