% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/languageContext.R
\name{isContextRunning}
\alias{isContextRunning}
\title{check if context is running}
\usage{
isContextRunning(s)
}
\arguments{
\item{ctx}{a context}
}
\description{
check if context is running
}

