% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/languageContext.R
\name{dbxRunCommand}
\alias{dbxRunCommand}
\title{Command execute}
\usage{
dbxRunCommand(command, ctx, wait = 0, language = "python",
  poll.log = TRUE, instance = options("databricks")[[1]]$instance,
  clusterId = options("databricks")[[1]]$clusterId,
  user = options("databricks")[[1]]$user,
  password = options("databricks")[[1]]$password)
}
\arguments{
\item{command}{a string command to be executed}

\item{ctx}{the context for the language}

\item{wait}{if non zero, will wait till command is finnished. wait is seconds polling}

\item{language}{is the language of the command}

\item{poll.log}{TRUE if we need to poll the s3 log file for output from remote code}

\item{instance}{is the instance of databricks}

\item{clusterId}{is the clusterId you're working with}

\item{user}{your usename}

\item{password}{your password}
}
\value{
a commandId
}
\description{
Command execute
}
\details{
see https://docs.databricks.com/api/1.2/index.html#command-execution
}

